added authorization to update+delete

+++ backend/routes/posts.js
@@ -56,6 +56,8 @@ router.post(
     });
 });
 
+
+// UPDATE
 router.put(
   "/:id",
   checkAuth,
@@ -73,11 +75,18 @@ router.put(
       imagePath: imagePath
     });
     Post
-      .updateOne({_id: req.params.id}, post)
+      .updateOne({_id: req.params.id, creator: req.params.userId}, post)
       .then(result => {
-        res.status(200).json({
-          message: "Update successful!"
-        });
+        if (result.n > 0) {
+          res.status(200).json({
+            message: "Update successful!"
+          });
+        } else {
+          res.status(401).json({
+            message: "Not authorized!"
+          });
+        }
+
       });
   }
 );
@@ -127,12 +136,17 @@ router.delete(
   checkAuth,
   (req, res, next) => {
     Post
-      .deleteOne({_id: req.params.id})
+      .deleteOne({_id: req.params.id, creator: req.userData.userId})
       .then(result => {
-        console.log(result);
-        res.status(200).json({
-          message: "Post deleted!"
-        });
+        if(result.n > 0) {
+          res.status(200).json({
+            message: "Deletion successful!"
+          });
+        } else {
+          res.status(401).json({
+            message: "Not authorized!"
+          });
+        }
       });
   }
 );
